backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
Pillow==10.1.0
pdf2image==1.16.3


# Temporal Cloud (will be added later)
# temporalio==1.5.0

# LLM (will be added later)
# openai==1.3.0
# anthropic==0.7.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# HTTP client
httpx==0.25.2

# Surya OCR (external service; no local OCR libs required)

